Commit for posterity: scss and css files
Once a new scss build was setup, I realized that main.css file was
heavily edited and the corresponding scss files were not. The subsequent
commit will pull the scss files.

This commit is intended to capture the state of the scss files and the
build setup before its removed.